backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
cryptography>=41.0.0
redis==5.0.1
prometheus-client==0.19.0
aiofiles==23.2.1